# packages/poster-contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Social Poster internal API",
    "version": "1.4",
    "description": "Cross-platform publishing API. Two auth modes share every /v1 route (D-023): app mode uses a client-credentials JWT, user mode uses the signed-in user’s Supabase session JWT and acts as the first-party poster-web app."
  },
  "servers": [
    {
      "url": "/",
      "description": "The service root; deployments front this with TLS."
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "App mode: the token from POST /v1/oauth/token. User mode: a Supabase session JWT."
      }
    },
    "schemas": {
...
      "SubmitPostRequest": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "external_ref": {
            "description": "The client’s own id, echoed on every webhook for this post.",
            "type": "string",
            "maxLength": 255
          },
          "content": {
            "type": "object",
            "properties": {
              "text": {
                "description": "Default caption for every target.",
                "type": "string"
              },
              "title": {
                "description": "Default title, for platforms that have one.",
                "type": "string"
              },
              "media": {
                "description": "Public media ids, in order.",
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "thread": {
                "description": "Ordered multi-part post (FR-08). Rejected on platforms without threads.",
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string"
                    },
                    "media": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": ["text"],
                  "additionalProperties": false
                }
              }
            },
            "additionalProperties": false,
            "description": "What to post, before per-target overrides."
          },
          "targets": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "connection_id": {
                  "type": "string",
                  "description": "Public connection id, e.g. cn_…"
                },
                "overrides": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false,
                  "description": "Per-platform replacements for the default content."
                }
              },
              "required": ["connection_id"],
              "additionalProperties": false,
              "description": "One target: which account, and what to change for it."
            }
          },
          "schedule_at": {
            "description": "ISO 8601 UTC. Omitted means dispatch as soon as the post is ready.",
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": ["user_id", "content", "targets"],
        "additionalProperties": false,
        "description": "A post to publish to one or more connected accounts."
      },
      "Post": {
        "type": "object",
        "properties": {
          "post_id": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "accepted",
              "scheduled",
              "dispatching",
              "posted",
              "failed",
              "paused",
              "canceled",
              "partial"
            ],
            "description": "Derived from the targets: posted when all posted, partial when mixed."
          },
          "external_ref": {
            "nullable": true,
            "type": "string"
          },
          "schedule_at": {
            "nullable": true,
            "type": "string"
          },
          "content": {
            "type": "object",
            "properties": {
              "text": {
                "description": "Default caption for every target.",
                "type": "string"
              },
              "title": {
                "description": "Default title, for platforms that have one.",
                "type": "string"
              },
              "media": {
                "description": "Public media ids, in order.",
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "thread": {
                "description": "Ordered multi-part post (FR-08). Rejected on platforms without threads.",
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string"
                    },
                    "media": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": ["text"],
                  "additionalProperties": false
                }
              }
            },
            "additionalProperties": false,
            "description": "What to post, before per-target overrides."
          },
          "created_at": {
            "type": "string"
          },
          "targets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "target_id": {
                  "type": "string"
                },
                "connection_id": {
                  "type": "string"
                },
                "platform_id": {
                  "type": "string"
                },
                "state": {
                  "type": "string",
                  "enum": [
                    "accepted",
                    "scheduled",
                    "dispatching",
                    "posted",
                    "failed",
                    "paused",
                    "canceled"
                  ],
                  "description": "Per-target state (§6). State lives per target, never per post."
                },
                "due_at": {
                  "type": "string"
                },
                "position": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991,
                  "description": "Matches target_index in §8 details."
                }
              },
              "required": [
                "target_id",
                "connection_id",
                "platform_id",
                "state",
                "due_at",
                "position"
              ],
              "additionalProperties": false,
              "description": "One created target."
            }
          }
        },
        "required": [
          "post_id",
          "state",
          "external_ref",
          "schedule_at",
          "content",
          "created_at",
          "targets"
        ],
        "additionalProperties": false,
        "description": "A logical post and its targets."
      },
...
      "PatchPostRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "object",
            "properties": {
              "text": {
                "description": "Default caption for every target.",
                "type": "string"
              },
              "title": {
                "description": "Default title, for platforms that have one.",
                "type": "string"
              },
              "media": {
                "description": "Public media ids, in order.",
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "thread": {
                "description": "Ordered multi-part post (FR-08). Rejected on platforms without threads.",
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string"
                    },
                    "media": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": ["text"],
                  "additionalProperties": false
                }
              }
            },
            "additionalProperties": false,
            "description": "What to post, before per-target overrides."
          },
          "targets": {
            "description": "Replaces the target list wholesale. Omit to leave targets alone.",
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "connection_id": {
                  "type": "string",
                  "description": "Public connection id, e.g. cn_…"
                },
                "overrides": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false,
                  "description": "Per-platform replacements for the default content."
                }
              },
              "required": ["connection_id"],
              "additionalProperties": false,
              "description": "One target: which account, and what to change for it."
            }
          },
          "schedule_at": {
            "description": "null clears the schedule.",
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "additionalProperties": false,
        "description": "An edit before dispatch. Re-validates constraints (FR-13)."
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "invalid_request",
                  "invalid_token",
                  "forbidden_user",
                  "grant_missing",
                  "not_found",
                  "idempotency_conflict",
                  "too_late",
                  "constraint_violation",
                  "rate_limited",
                  "internal_error"
                ],
                "description": "Machine-readable error code."
              },
              "message": {
                "type": "string",
                "description": "Human-readable summary. Never used for control flow."
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "target_index": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991,
                      "description": "Index of the failing target in the submitted targets array."
                    },
                    "connection_id": {
                      "description": "Public id of the target connection.",
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "description": "Constraint code; see the contract §8 list."
                    },
                    "constraint": {
                      "description": "The limit that was exceeded and the actual value, as data.",
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": ["target_index", "code"],
                  "additionalProperties": false,
                  "description": "One per failing target, so a client can fix exactly what broke."
                }
              },
              "request_id": {
                "type": "string",
                "description": "Echoes the X-Request-Id response header. Quote it in support requests."
              }
            },
            "required": ["code", "message", "request_id"],
            "additionalProperties": false
          }
        },
        "required": ["error"],
        "additionalProperties": false,
        "description": "The single error envelope used by every non-2xx response."
      }
    }
  },
  "paths": {
...
    "/v1/posts": {
      "post": {
        "operationId": "submitPost",
        "summary": "Submit a post to one or more connected accounts",
        "description": "Validates every target and creates all of them or none (§5). Send an `Idempotency-Key` header: the same key with the same body returns the original result, and the same key with a different body is 409 `idempotency_conflict`. Keys are scoped per app and kept at least 24 hours. Omitting `schedule_at` means dispatch as soon as the post is ready.",
        "tags": ["posts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "Strongly recommended. Without it a retry creates a second post."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitPostRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted; targets created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or expired token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "forbidden_user, or grant_missing for a connection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "A connection or media id does not exist for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Same Idempotency-Key with a different body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "One or more targets failed platform rules; see details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-app rate limit exceeded; honor Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{post_id}": {
      "get": {
        "operationId": "getPost",
        "summary": "Read a post and its targets",
        "tags": ["posts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The post, with per-target state and outcomes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetail"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or expired token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such post for this app and user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patchPost",
        "summary": "Edit a post before dispatch",
        "description": "Re-validates constraints (FR-13). Refused with 409 `too_late` once any target has begun dispatching, because the content may already be on its way.",
        "tags": ["posts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchPostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated and re-validated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetail"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or expired token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "forbidden_user, or grant_missing for a connection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such post, connection, or media for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "A target is already dispatching.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "The edit fails platform rules; see details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{post_id}/cancel": {
      "post": {
        "operationId": "cancelPost",
        "summary": "Cancel a post before dispatch",
        "description": "Cancels every target that has not begun dispatching (§6). Returns 409 `too_late` when nothing could be canceled. A post already canceled returns 200 with an empty `canceled_target_ids`, so a retry is safe.",
        "tags": ["posts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancel applied, or already canceled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelPostResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or expired token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such post for this app and user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Every target had already begun dispatching or finished.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  }
}